feat: add domain-specific firestore helper modules
