+++ helpers/wishes.ts
@@ -18,19 +18,12 @@ import {
 } from 'firebase/firestore';
 import { db } from '../firebase';
 import type { Wish } from '../types/Wish';
+import { getFollowingIds } from './followers';
 
-export interface Comment {
-  id: string;
-  text: string;
-  userId?: string;
-  displayName?: string;
-  photoURL?: string;
-  isAnonymous?: boolean;
-  timestamp?: any;
-  parentId?: string;
-  reactions?: Record<string, number>;
-  userReactions?: Record<string, string>;
-  [key: string]: any;
+export interface TopCreator {
+  userId: string;
+  displayName: string;
+  count: number;
 }
 
 export function listenTrendingWishes(cb: (wishes: Wish[]) => void) {
@@ -118,75 +111,9 @@ export function listenBoostedWishes(cb: (wishes: Wish[]) => void) {
   });
 }
 
-export async function followUser(currentUser: string, targetUser: string) {
-  const followerRef = doc(db, 'users', targetUser, 'followers', currentUser);
-  const followingRef = doc(db, 'users', currentUser, 'following', targetUser);
-  await Promise.all([
-    setDoc(followerRef, { createdAt: serverTimestamp() }),
-    setDoc(followingRef, { createdAt: serverTimestamp() }),
-  ]);
-}
-
-export async function unfollowUser(currentUser: string, targetUser: string) {
-  const followerRef = doc(db, 'users', targetUser, 'followers', currentUser);
-  const followingRef = doc(db, 'users', currentUser, 'following', targetUser);
-  await Promise.all([deleteDoc(followerRef), deleteDoc(followingRef)]);
-}
-
-export async function getFollowingIds(userId: string): Promise<string[]> {
-  const q = query(collection(db, 'users', userId, 'following'));
-  const snap = await getDocs(q);
-  return snap.docs.map((d) => d.id);
-}
-
-export function listenFollowingWishes(
-  userId: string,
-  cb: (wishes: Wish[]) => void,
-) {
-  let unsub = () => {};
-  getFollowingIds(userId).then((ids) => {
-    if (ids.length === 0) {
-      cb([]);
-      return;
-    }
-    const q = query(
-      collection(db, 'wishes'),
-      where('userId', 'in', ids),
-      orderBy('timestamp', 'desc'),
-    );
-    unsub = onSnapshot(q, (snap) => {
-      const data = snap.docs.map((d) => ({
-        id: d.id,
-        ...(d.data() as Omit<Wish, 'id'>),
-      }));
-      cb(data as Wish[]);
-    });
-  });
-  return () => unsub();
-}
-
-export async function getFollowingWishes(userId: string): Promise<Wish[]> {
-  const ids = await getFollowingIds(userId);
-  if (ids.length === 0) return [];
-  const q = query(
-    collection(db, 'wishes'),
-    where('userId', 'in', ids),
-    orderBy('timestamp', 'desc'),
-  );
-  const snap = await getDocs(q);
-  return snap.docs.map((d) => ({
-    id: d.id,
-    ...(d.data() as Omit<Wish, 'id'>),
-  })) as Wish[];
-}
-
-export async function getTopBoostedCreators(limitCount = 5): Promise<
-  {
-    userId: string;
-    displayName: string;
-    count: number;
-  }[]
-> {
+export async function getTopBoostedCreators(
+  limitCount = 5,
+): Promise<TopCreator[]> {
   const since = new Date(Date.now() - 7 * 24 * 60 * 60 * 1000);
   const q = query(collection(db, 'wishes'), where('boostedUntil', '>=', since));
   const snap = await getDocs(q);
@@ -225,8 +152,7 @@ export async function getWhispOfTheDay(): Promise<Wish | null> {
       w.boostedUntil.toDate() > new Date();
     const reacts =
       !!w.reactions &&
-      Object.values(w.reactions).reduce<number>((sum, v) => sum + (v ?? 0), 0) >
-        0;
+      Object.values(w.reactions).reduce<number>((sum, v) => sum + (v ?? 0), 0) > 0;
     return boost || reacts;
   });
   if (filtered.length === 0) return null;
@@ -317,62 +243,6 @@ export async function getWish(id: string): Promise<Wish | null> {
     : null;
 }
 
-export function listenWishComments(
-  wishId: string,
-  cb: (comments: Comment[]) => void,
-) {
-  const q = query(
-    collection(db, 'wishes', wishId, 'comments'),
-    orderBy('timestamp', 'asc'),
-  );
-  return onSnapshot(q, (snap) => {
-    const data = snap.docs.map((d) => ({
-      id: d.id,
-      ...(d.data() as Omit<Comment, 'id'>),
-    }));
-    cb(data as Comment[]);
-  });
-}
-
-export async function addComment(wishId: string, data: Omit<Comment, 'id'>) {
-  return addDoc(collection(db, 'wishes', wishId, 'comments'), {
-    timestamp: serverTimestamp(),
-    ...data,
-  });
-}
-
-export async function updateCommentReaction(
-  wishId: string,
-  commentId: string,
-  emoji: string,
-  prevEmoji: string | undefined,
-  user: string,
-) {
-  const ref = doc(db, 'wishes', wishId, 'comments', commentId);
-  const snap = await getDoc(ref);
-  if (!snap.exists()) return;
-  const data = snap.data() as Comment;
-  const reactions = { ...(data.reactions || {}) } as Record<string, number>;
-  const userReactions = { ...(data.userReactions || {}) } as Record<
-    string,
-    string
-  >;
-
-  if (prevEmoji && reactions[prevEmoji]) {
-    reactions[prevEmoji] -= 1;
-    if (reactions[prevEmoji] === 0) delete reactions[prevEmoji];
-  }
-
-  if (prevEmoji === emoji) {
-    delete userReactions[user];
-  } else {
-    userReactions[user] = emoji;
-    reactions[emoji] = (reactions[emoji] || 0) + 1;
-  }
-
-  return updateDoc(ref, { reactions, userReactions });
-}
-
 export async function getWishesByNickname(nickname: string): Promise<Wish[]> {
   const snap = await getDocs(
     query(collection(db, 'wishes'), where('nickname', '==', nickname)),
@@ -459,26 +329,10 @@ export async function getAllWishes(): Promise<Wish[]> {
   return wishes;
 }
 
-export async function getWishComments(wishId: string): Promise<Comment[]> {
-  const snap = await getDocs(collection(db, 'wishes', wishId, 'comments'));
-  return snap.docs.map((d) => {
-    const data = d.data();
-    const comment: Comment = {
-      id: d.id,
-      text: data.text,
-      nickname: data.nickname,
-      timestamp: data.timestamp,
-      parentId: data.parentId,
-      reactions: data.reactions,
-      userReactions: data.userReactions,
-    };
-    return comment;
-  });
-}
-
 export async function cleanupExpiredWishes() {
   const now = new Date();
   const q = query(collection(db, 'wishes'), where('expiresAt', '<=', now));
   const snap = await getDocs(q);
   await Promise.all(snap.docs.map((d) => deleteDoc(d.ref)));
 }
+
